+++ assets/js/helpscout-beacon.js
@@ -2,7 +2,7 @@
 * @Author: Timi Wahalahti
 * @Date:   2020-02-12 14:35:14
 * @Last Modified by:   Timi Wahalahti
-* @Last Modified time: 2020-02-12 15:26:35
+* @Last Modified time: 2022-11-09 14:52:02
 */
 
 !function(e,t,n){function a(){var e=t.getElementsByTagName("script")[0],n=t.createElement("script");n.type="text/javascript",n.async=!0,n.src="https://beacon-v2.helpscout.net",e.parentNode.insertBefore(n,e)}if(e.Beacon=n=function(t,n,a){e.Beacon.readyQueue.push({method:t,options:n,data:a})},n.readyQueue=[],"complete"===t.readyState)return a();e.attachEvent?e.attachEvent("onload",a):e.addEventListener("load",a,!1)}(window,document,window.Beacon||function(){});