+++ inc/admin/helpscout.php
@@ -3,7 +3,7 @@
  * @Author: Timi Wahalahti
  * @Date:   2020-02-12 14:29:27
  * @Last Modified by:   Timi Wahalahti
- * @Last Modified time: 2021-08-02 14:24:25
+ * @Last Modified time: 2022-11-09 14:54:24
  *
  * @package air-helper
  */
@@ -45,6 +45,7 @@ function air_helper_enqueue_helpscout_beacon() {
   // Increase body padding for the HS widget not to override paging controls
   add_action( 'admin_head', function() { ?>
     <style type="text/css">
+      .block-editor .edit-post-sidebar,
       #wpbody-content {
         padding-bottom: 100px;
       }