+++ air-helper.php
@@ -7,7 +7,7 @@
  * Author: Digitoimisto Dude Oy, Timi Wahalahti
  * Author URI: https://www.dude.fi
  * Requires at least: 5.5
- * Tested up to: 5.0
+ * Tested up to: 6.1
  * License: GPL-3.0+
  * License URI: https://www.gnu.org/licenses/gpl-3.0.html
  *