+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ## Unreleased
 ### Added
 - Dismiss Filebird version 5.0.8 YayMail plugin upsell
+- Disable autoload for notification dismissal options
 
 ## [2.16.2]
 ### Changed